+++ tests/test_checker/test_noqa.py
@@ -99,6 +99,7 @@ def test_noqa_fixture_disabled(absolute_path, all_violations):
         'Z327': 1,
         'Z328': 2,
         'Z329': 1,
+        'Z330': 1,
 
         'Z400': 0,
         'Z401': 0,