+++ CHANGELOG.md
@@ -4,6 +4,16 @@ We follow Semantic Versions since the `0.1.0` release.
 We used to have incremental versioning before `0.1.0`.
 
 
+
+## 0.8.0
+
+
+### Features
+
+- Forbids to use redundant `+` operator before numbers
+- Forbids to use complex default values
+
+
 ## 0.7.1
 
 ### Bugfixes